Rename board files to lowercase
